==> 1_load_data_2023_10_02.ipynb <==
# Creating a data table from Arivale for prediction

!ls ../copied_csvs

## 1. Load data, preprocessing

import numpy as np
import pandas as pd

chems = pd.read_csv('../copied_csvs/chems_all.csv', index_col=0, dtype={'public_client_id': np.str_})
clients = pd.read_csv('../copied_csvs/clients.csv', index_col=0, dtype={'public_client_id': np.str_})
bp = pd.read_csv('../copied_csvs/bp.csv', index_col=0, dtype={'public_client_id': np.str_})
weight = pd.read_csv('../copied_csvs/weight.csv', index_col=0, dtype={'public_client_id': np.str_})
meds = pd.read_csv('../copied_csvs/meds.csv', index_col=0, dtype={'public_client_id': np.str_})

prots_t2d = pd.read_csv('../copied_csvs/prots_t2d.csv', index_col=0,
                       dtype={'public_client_id':np.str_})
prots_not_t2d = pd.read_csv('../copied_csvs/prots_not_t2d.csv', index_col=0,
                            dtype={'public_client_id':np.str_})
prots_all = pd.concat([prots_t2d, prots_not_t2d])
# we want to keep at least one is_t2d column.

mets_t2d = pd.read_csv('../copied_csvs/mets_t2d.csv', index_col=0,
                       dtype={'public_client_id':np.str_})
mets_not_t2d = pd.read_csv('../copied_csvs/mets_not_t2d.csv', index_col=0,
                            dtype={'public_client_id':np.str_})
mets_all = pd.concat([mets_t2d, mets_not_t2d])
if 'is_t2d' in mets_all.columns:
    del mets_all['is_t2d']

samples_t2d = pd.read_csv('../copied_csvs/t2d_samples.csv', index_col=0,
                       dtype={'public_client_id':np.str_})
samples_t2d = pd.read_csv('../copied_csvs/not_t2d_samples.csv', index_col=0,
                            dtype={'public_client_id':np.str_})
samples_all = pd.concat([samples_t2d, samples_t2d])

### Analysis of questionnaires and medications

diabetes_type2_current = 'assessment:health-history:cardiometabolic_type2_diabetes:self_current'
samples_all[diabetes_type2_current]
samples_all['is_t2d'] = samples_all[diabetes_type2_current]

samples_all['is_t1d'] = samples_all['assessment:health-history:autoimmune_type1_diabetes:self_current']

samples_all['assessment:health-history:cancer_basal_cell_carcinoma:self_current']
samples_all['assessment:health-history:cardiometabolic_metabolic_syndrome:self_current']
samples_all['assessment:health-history:cardiovascular_hypertension:self_current']
samples_all['assessment:health-history:bladder_kidney_kidney_disease:self_current']

samples_all.columns[['med' in x for x in samples_all.columns]]

samples_all['meds_blood_sugar'].unique()

samples_all['meds_blood_pressure'].unique()

### Removing duplicates

chems = chems.sort_values(['public_client_id', 'days_in_program'])
print(chems.shape)

bp.index = pd.MultiIndex.from_frame(bp[['public_client_id', 'days_in_program']])
bp = bp[~bp.index.duplicated(keep='first')]

chems.index = pd.MultiIndex.from_frame(chems[['public_client_id', 'days_in_program']])
chems = chems[~chems.index.duplicated(keep='first')]

weight.index = pd.MultiIndex.from_frame(weight[['public_client_id', 'days_in_program']])
weight = weight[~weight.index.duplicated(keep='first')]

prots_all.index = pd.MultiIndex.from_frame(prots_all[['public_client_id', 'days_in_program']])
prots_all = prots_all[~prots_all.index.duplicated(keep='first')]

mets_all.index = pd.MultiIndex.from_frame(mets_all[['public_client_id', 'days_in_program']])
mets_all = mets_all[~mets_all.index.duplicated(keep='first')]

samples_all.index = pd.MultiIndex.from_frame(samples_all[['public_client_id', 'days_in_program']])
samples_all = samples_all[~samples_all.index.duplicated(keep='first')]

print(chems.shape)

### Generating a prev/next field for chems

chems['days_since_previous'] = [0]*len(chems)
chems['days_til_next'] = [0]*len(chems)

chems_selected = ['HbA1C', 'GFR', 'GLUCOSE', 'INSULIN', 'HOMA-IR']
chems_to_column = {
    'HbA1C': 'GLYCOHEMOGLOBIN A1C',
    'GFR': 'GFR, MDRD',
    'GLUCOSE': 'GLUCOSE',
    'INSULIN': 'INSULIN',
    'HOMA-IR': 'HOMA-IR'
}

for c in chems_selected:
    chems['d_'+c] = [0]*len(chems)
    chems['prev_'+c] = [0]*len(chems)
    chems['next_'+c] = [0]*len(chems)

current_id = None
prev_time = 0
prev_chems = {c: 0 for c in chems_selected}
prev_glucose = 0
prev_hba1c = 0
prev_gfr = 0
for i, (index, row) in enumerate(chems.iterrows()):
    if row.public_client_id != current_id:
        current_id = row.public_client_id
    else:
        chems.loc[index, 'days_since_previous'] = row.days_in_program - prev_time
        for c in chems_selected:
            chems.loc[index, 'prev_' + c] = prev_chems[c]
    prev_time = row.days_in_program
    for c in chems_selected:
        prev_chems[c] = row[chems_to_column[c]]
    if i < len(chems) - 1:
        next_row = chems.iloc[i+1]
        if next_row.public_client_id == current_id:
            chems.loc[index, 'days_til_next'] = next_row.days_in_program - prev_time
            for c in chems_selected:
                chems.loc[index, 'next_' + c] = next_row[chems_to_column[c]]
                chems.loc[index, 'd_' + c] = next_row[chems_to_column[c]] - row[chems_to_column[c]]

#### Generating longer-term differences (1 year, 2 years, etc)

chems['days_til_next_1y'] = [0]*len(chems)
chems['days_til_next_2y'] = [0]*len(chems)

for c in chems_selected:
    chems['next_1y_'+c] = [np.nan]*len(chems)
    chems['d_1y_'+c] = [np.nan]*len(chems)
    chems['next_2y_'+c] = [np.nan]*len(chems)
    chems['d_2y_'+c] = [np.nan]*len(chems)

current_id = None
prev_time = 0
for i, (index, row) in enumerate(chems.iterrows()):
    if row.public_client_id != current_id:
        current_id = row.public_client_id
    prev_time = row.days_in_program
    j = i + 1
    while j < len(chems) - 1:
        next_row = chems.iloc[j]
        if next_row.public_client_id == current_id:
            days_til_next = next_row.days_in_program - prev_time
            if days_til_next < 420 and days_til_next > 300:
                chems.loc[index, 'days_til_next_1y'] = days_til_next
                for c in chems_selected:
                    chems.loc[index, 'next_1y_' + c] = next_row[chems_to_column[c]]
                    chems.loc[index, 'd_1y_' + c] = next_row[chems_to_column[c]] - row[chems_to_column[c]]
            if days_til_next > 660 and days_til_next < 780:
                chems.loc[index, 'days_til_next_2y'] = days_til_next
                for c in chems_selected:
                    chems.loc[index, 'next_2y_' + c] = next_row[chems_to_column[c]]
                    chems.loc[index, 'd_2y_' + c] = next_row[chems_to_column[c]] - row[chems_to_column[c]]
            j += 1
        else:
            break

print((~chems['next_1y_GFR'].isna()).sum(), (~chems['next_2y_GFR'].isna()).sum())

## 2. Creating a combined clinical-metabolomic-proteomic table

combined_table_mets = pd.concat([mets_all, prots_all, chems], axis=1)

combined_table_mets_filtered = combined_table_mets[~combined_table_mets['public_client_id'].isna().any(1)]

combined_table_mets_filtered = combined_table_mets_filtered.loc[:, ~combined_table_mets_filtered.columns.duplicated()]

combined_table_mets_filtered.shape

# TODO: this messes up the dtypes
combined_table_mets_filtered = combined_table_mets_filtered.T.drop_duplicates(keep='first').T
combined_table_mets_filtered = combined_table_mets_filtered.infer_objects()

combined_table_mets_filtered.shape

combined_table_mets_filtered = combined_table_mets_filtered.sort_index()

### Adding additional fields to the table - age, weight, meds, etc

# TODO: add meds_blood_sugar as a feature
samples_all.index = samples_all.public_client_id
weight.index = weight.public_client_id

def get_nearest_index(df, col, val):
    "Gets the nearest index to the val, given a df sorted by col."
    min_distance = np.inf
    best_index = None
    best_row = None
    for index, row in df.iterrows():
        distance = np.abs(row[col] - val)
        if distance < min_distance:
            min_distance = distance
            best_index = index
            best_row = row
        else:
            if best_index is not None:
                return best_index, best_row
    return best_index, best_row

clients.index = clients.public_client_id
combined_table_mets_filtered = combined_table_mets_filtered.copy()
combined_table_mets_filtered['age'] = np.zeros(combined_table_mets_filtered.shape[0]) + np.nan
combined_table_mets_filtered['is_m'] = np.zeros(combined_table_mets_filtered.shape[0]) + np.nan
combined_table_mets_filtered['weight'] = np.zeros(combined_table_mets_filtered.shape[0]) + np.nan
combined_table_mets_filtered['height'] = np.zeros(combined_table_mets_filtered.shape[0]) + np.nan
combined_table_mets_filtered['bmi'] = np.zeros(combined_table_mets_filtered.shape[0]) + np.nan
combined_table_mets_filtered['meds_blood_sugar'] = np.zeros(combined_table_mets_filtered.shape[0]) + np.nan


for i, row in combined_table_mets_filtered.iterrows():
    # ci is client id
    ci = i[0]
    days = i[1]
    try:
        combined_table_mets_filtered.loc[i,'age'] = clients.loc[ci].age
        combined_table_mets_filtered.loc[i,'is_m'] = int(clients.loc[ci].sex == 'M')
    except:
        print('age not available for', ci)
    try:
        weights_subset = weight.loc[ci]
        if isinstance(weights_subset, pd.DataFrame):
            weight_index, weight_row = get_nearest_index(weights_subset, 'days_in_program', days)
        else:
            weight_row = weights_subset
        combined_table_mets_filtered.loc[i,'weight'] = weight_row.WEIGHT_CALC
        combined_table_mets_filtered.loc[i,'height'] = weight_row.HEIGHT_CALC
        combined_table_mets_filtered.loc[i,'bmi'] = weight_row.BMI_CALC
    except:
        print('weights not available for', ci)
    try:
        samples_subset = samples_all.loc[ci]
        if isinstance(samples_subset, pd.DataFrame):
            sample_index, sample_row = get_nearest_index(samples_subset, 'days_in_program', days)
            combined_table_mets_filtered.loc[i,'meds_blood_sugar'] = sample_row.meds_blood_sugar
        else:
            combined_table_mets_filtered.loc[i,'meds_blood_sugar'] = samples_subset.meds_blood_sugar
    except:
        print('sample questionnaires not available for', ci)

(combined_table_mets_filtered.next_HbA1C == 0).sum()

(combined_table_mets_filtered.next_HbA1C.isna()).sum()

(combined_table_mets_filtered.next_GLUCOSE.isna()).sum()

### Averaging duplicated protein columns - if protein columns are duplicated across sets, average them.

prot_names = pd.read_csv('../arivale_data/arivale_prots.tsv', sep='\t')
gene_names = prot_names.groupby('gene_name')['index'].unique()
duplicate_gene_names = gene_names[gene_names.map(len) > 1]
combined_table_mets_prots = combined_table_mets_filtered.copy()
for index, ids in duplicate_gene_names.items():
    table_ids = [i for i in combined_table_mets_filtered.columns if i in ids]
    if len(table_ids) > 1:
        print(index, table_ids)
        print(combined_table_mets_filtered[table_ids].corr())
        table_mean = combined_table_mets_filtered[table_ids].mean(axis=1, skipna=True)
        combined_table_mets_prots[table_ids[0]] = table_mean
        combined_table_mets_prots = combined_table_mets_prots.drop(table_ids[1:], axis=1)

print(combined_table_mets_prots.shape)

### Taking only the first time point

#  try taking only the first patient?
first_items = []
prev_client = None
for c in combined_table_mets_prots.public_client_id:
    if c == prev_client:
        first_items.append(False)
    else:
        prev_client = c
        first_items.append(True)
first_items = np.array(first_items)

combined_table_mets_filtered_first_items = combined_table_mets_prots[first_items]

print(combined_table_mets_filtered_first_items.shape)

### Filtering columns by NaNs, selecting columns to use

1. remove rows that have missing next-time-point values
2. identify -omics/clinical columns that have less than 10% missingness, select these columns to use for further analysis

combined_table_mets_filtered_1 = combined_table_mets_filtered_first_items[(combined_table_mets_filtered_first_items.next_HbA1C != 0) &\
                                                (combined_table_mets_filtered_first_items.next_GFR != 0) &\
                                                (combined_table_mets_filtered_first_items.next_GLUCOSE != 0) &\
                                                (combined_table_mets_filtered_first_items.next_INSULIN != 0) &\
                                                (~combined_table_mets_filtered_first_items.next_HbA1C.isna()) &\
                                                (~combined_table_mets_filtered_first_items.next_GFR.isna()) &\
                                                (~combined_table_mets_filtered_first_items.next_GLUCOSE.isna()) &\
                                                (~combined_table_mets_filtered_first_items.next_INSULIN.isna())]

for chem, col_name in chems_to_column.items():
    combined_table_mets_filtered_1 = combined_table_mets_filtered_1[~combined_table_mets_filtered_1[col_name].isna()]

print(combined_table_mets_filtered_1.shape)

na_counts = combined_table_mets_filtered_1.isna().sum(0)

Cutofff of 10% missingness for columns

less_than_10 = (na_counts < combined_table_mets_filtered_1.shape[0]*0.1) & (combined_table_mets_filtered_1.dtypes == np.float64)

is_next = combined_table_mets_filtered_1.columns.str.contains('d_') |\
          combined_table_mets_filtered_1.columns.str.contains('next_')| \
          combined_table_mets_filtered_1.columns.str.contains('prev_')

selected_columns_full = combined_table_mets_filtered_1.columns[less_than_10 & ~is_next].to_numpy()
columns_to_exclude = set(['client_id', 'days_til_next', 'days_since_previous', 'days_til_next_1y', 'days_til_next_2y',
                          'observation_id', 'days_in_program',
                          'days_since_first_call', 'days_since_first_draw'])
selected_columns_full = np.array([x for x in selected_columns_full if x not in columns_to_exclude])

selected_chem_bp_cols = [x for x in selected_columns_full if x in chems.columns or x in bp.columns or x in ['age', 'is_m', 'bmi', 'meds_blood_sugar'] or x in weight.columns]

print('number of total clinical columns:', len(selected_chem_bp_cols))

print(selected_chem_bp_cols)

selected_chems_only = [x for x in selected_columns_full if x in chems.columns]


chem_subset_cols = ['INSULIN', 'GLUCOSE', 'GLYCOHEMOGLOBIN A1C', 'GFR, MDRD',
                  'ALBUMIN', 'CREATININE ENZ, SER',
                  'HDL CHOL DIRECT', 'HEMATOCRIT', 'HEMOGLOBIN', 'LDL-CHOL CALCULATION', 'TRIGLYCERIDES',
                  'MEAN_ARTERIAL_BLOOD_PRESSURE', 'PULSE_PRESSURE', 'diastolic', 'systolic',
                  'age', 'is_m', 'meds_blood_sugar', 'bmi']
chem_subset_cols = [x for x in chem_subset_cols if x in selected_columns_full]
print('number of selected clinical columns:', len(chem_subset_cols))
print(chem_subset_cols)

Somewhat arbitrarily, we're only including proteins with less than 5 missing values. This is different from clinical variables because of the different distributions of missing-ness; having a more stringent missingness threshold will remove a lot of clinical variables that we want to include, and will also remove too many metabolites.

less_than_5 = (na_counts < 5)

selected_prot_cols = [x for x in selected_columns_full if x in prots_all.columns and less_than_5[x]]
print('number of protein columns:', len(selected_prot_cols))

selected_met_cols = [x for x in selected_columns_full if x in mets_all.columns]
print('number of metabolite columns:', len(selected_met_cols))

selected_columns_full = selected_chem_bp_cols + selected_prot_cols + selected_met_cols

print('number of total columns:', len(selected_columns_full))

### Remove rows that have NaN values for chems, set metabolites with NaNs to 0

try: setting missing metabolites to the median, remove rows with missing chems.

combined_table_mets_filtered_1['bmi'].isna().sum()

chem_has_na = combined_table_mets_filtered_1[selected_chem_bp_cols].isna().sum(1)
met_has_na = combined_table_mets_filtered_1[selected_met_cols].isna().sum(1)
prot_has_na = combined_table_mets_filtered_1[selected_prot_cols].isna().sum(1)
has_na = chem_has_na + met_has_na + prot_has_na

nans_above_threshold = (has_na > combined_table_mets_filtered_1.shape[0]*0.1)
print('# of samples with nans > 10%:', nans_above_threshold.sum())

print('# of samples with missing clinical var:', sum(chem_has_na > 0))

print('# of samples with missing protein:', sum(prot_has_na > 0))

# only remove the chem and prot NaN rows, set the met NaN rows to 0
combined_table_mets_filtered_2 = combined_table_mets_filtered_1[(~nans_above_threshold) & (chem_has_na == 0) & (prot_has_na == 0)]

combined_table_mets_set_medians = combined_table_mets_filtered_2.copy()


print('shape after removing values with missing chems/prots:', combined_table_mets_filtered_2.shape)

### Filter by time interval between samples

combined_table_mets_filtered_times = combined_table_mets_filtered_2[\
                                                                (combined_table_mets_filtered_2.days_til_next < 270) \
                                                                & (combined_table_mets_filtered_2.days_til_next > 90)]
print(combined_table_mets_filtered_times.shape)

combined_table_mets_set_zeros = combined_table_mets_filtered_times.copy()
combined_table_mets_set_zeros[selected_met_cols] = combined_table_mets_set_zeros[selected_met_cols].fillna(0)
combined_table_mets_set_zeros[selected_prot_cols] = combined_table_mets_set_zeros[selected_prot_cols].fillna(0)


### Imputation for metabolites using missforest

import os
from missingpy import MissForest
import warnings
warnings.filterwarnings("ignore")

imputer = MissForest(max_iter=5, decreasing=False, missing_values=np.nan,
                     copy=True, criterion=('squared_error', 'gini'),
                     max_depth=15, min_samples_split=2, min_samples_leaf=1,
                     min_weight_fraction_leaf=0.0,
                     max_leaf_nodes=None, min_impurity_decrease=0.0,
                     bootstrap=True, oob_score=False, n_jobs=-1,
                     verbose=0, warm_start=True, class_weight=None,
                     max_features=1)

data_selected_cols = combined_table_mets_filtered_times[selected_met_cols]

imputed = imputer.fit_transform(data_selected_cols)

combined_table_mets_imputed = combined_table_mets_filtered_times.copy()

#if not os.path.exists('clinical_multiomics_combined_data_imputed.csv'):
data_selected_cols[:] = imputed
#else:
#    data_selected_cols = pd.read_csv('clinical_multiomics_combined_data_imputed.csv')
combined_table_mets_imputed[selected_met_cols] = data_selected_cols.values

combined_table_mets_imputed.shape

### Filter rows by output outliers

Thresholds are somewhat arbitrary, based on inspection

# TODO: remove outliers for next_hba1c, glucose, gfr
row_has_na = combined_table_mets_set_zeros[selected_met_cols].isna().any(1)
items_to_exclude = {'HbA1C', 'GLUCOSE', 'GFR', 'INSULIN'}
upper_bounds = {'HbA1C': 12, 'GLUCOSE': 200, 'GFR': 200, 'INSULIN': 60}
lower_bounds = {'HbA1C': 0, 'GLUCOSE': 0, 'GFR': 0, 'INSULIN': 0}

for k in items_to_exclude:
    item_data = combined_table_mets_imputed[chems_to_column[k]]
    mean = np.mean(item_data)
    std = np.std(item_data)
    row_has_na = row_has_na | (item_data > upper_bounds[k]) | (item_data < lower_bounds[k])
    item_data = combined_table_mets_set_medians['next_' + k]
    mean = np.mean(item_data)
    std = np.std(item_data)
    row_has_na = row_has_na | (item_data > upper_bounds[k]) | (item_data < lower_bounds[k])

combined_table_imputed_outliers = combined_table_mets_imputed[~row_has_na]
combined_table_mets_zeros_times_outliers = combined_table_mets_set_zeros[~row_has_na]

combined_table_imputed_outliers.shape

### Create data arrays for training/stats

#  try taking only the first patient?
first_items = []
prev_client = None
for c in combined_table_imputed_outliers.public_client_id:
    if c == prev_client:
        first_items.append(False)
    else:
        prev_client = c
        first_items.append(True)
first_items = np.array(first_items)

combined_table_imputed_outliers = combined_table_imputed_outliers[first_items]
combined_table_mets_zeros_times_outliers = combined_table_mets_zeros_times_outliers[first_items]

print(combined_table_imputed_outliers.shape)

#### Convert deltas to discrete values for classification

for key, val in chems_to_column.items():
    if key != 'GFR':
        pos = (combined_table_imputed_outliers['d_'+key] >= combined_table_imputed_outliers[val]*0.05)
        neg = (combined_table_imputed_outliers['d_'+key] < combined_table_imputed_outliers[val]*0.05)
    else:
        pos = (combined_table_imputed_outliers['d_'+key] <= combined_table_imputed_outliers[val]*-0.05)
        neg = (combined_table_imputed_outliers['d_'+key] > combined_table_imputed_outliers[val]*-0.05)
    combined_table_imputed_outliers['d_' + key + '_class'] = np.zeros(pos.shape) + pos
    combined_table_imputed_outliers['d_' + key + '_class'][combined_table_imputed_outliers['d_'+key].isna()] = pd.NA

for key, val in chems_to_column.items():
    if key != 'GFR':
        pos = (combined_table_imputed_outliers['d_1y_'+key] >= combined_table_imputed_outliers[val]*0.05)
        neg = (combined_table_imputed_outliers['d_1y_'+key] < combined_table_imputed_outliers[val]*0.05)
    else:
        pos = (combined_table_imputed_outliers['d_1y_'+key] <= combined_table_imputed_outliers[val]*-0.05)
        neg = (combined_table_imputed_outliers['d_1y_'+key] > combined_table_imputed_outliers[val]*-0.05)
    combined_table_imputed_outliers['d_1y_' + key + '_class'] = np.zeros(pos.shape) + pos
    combined_table_imputed_outliers['d_1y_' + key + '_class'][combined_table_imputed_outliers['d_1y_'+key].isna()] = pd.NA

## Saving table, saving variable lists

!mkdir results_2023_10_18

combined_table_imputed_outliers.to_csv('results_2023_10_18/combined_data_table.csv', index=None)

# save lists of chems, prots, mets
np.savetxt('results_2023_10_18/selected_columns_full.txt', selected_columns_full, fmt='%s')
np.savetxt('results_2023_10_18/chem_subset_cols.txt', chem_subset_cols, fmt='%s')
np.savetxt('results_2023_10_18/selected_chem_bp_cols.txt', selected_chem_bp_cols, fmt='%s')
np.savetxt('results_2023_10_18/selected_prot_cols.txt', selected_prot_cols, fmt='%s')
np.savetxt('results_2023_10_18/selected_met_cols.txt', selected_met_cols, fmt='%s')


## Aggregate data statistics

combined_table_imputed_outliers.agg({'age': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'bmi': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'GLYCOHEMOGLOBIN A1C': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'GLUCOSE': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'GFR, MDRD': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'INSULIN': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'HOMA-IR': ['count', 'min', 'mean', 'median', 'max', 'std'],}).T.style.format(precision=2)

combined_table_imputed_outliers.agg({'d_HbA1C': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'd_GLUCOSE': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'd_GFR': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'd_INSULIN': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'd_HOMA-IR': ['count', 'min', 'mean', 'median', 'max', 'std'],}).T.style.format(precision=2)

combined_table_imputed_outliers.agg({'d_1y_HbA1C': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'd_1y_GLUCOSE': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'd_1y_GFR': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'd_1y_INSULIN': ['count', 'min', 'mean', 'median', 'max', 'std'],
                    'd_1y_HOMA-IR': ['count', 'min', 'mean', 'median', 'max', 'std'],}).T.style.format(precision=2)
