==> 2_diffexp_2023_10_02.ipynb <==
# Differential Expression for clinical variables related to T2D

import pandas as pd
import numpy as np

## 1. Load data

combined_table = pd.read_csv('results_2023_10_18/combined_data_table.csv', dtype={'public_client_id': str})

combined_table.index = pd.MultiIndex.from_frame(combined_table[['public_client_id', 'days_in_program']])

selected_columns_full = np.loadtxt('results_2023_10_18/selected_columns_full.txt', dtype=str, delimiter='\t')
chem_subset_cols = np.loadtxt('results_2023_10_18/chem_subset_cols.txt', dtype=str, delimiter='\t')
selected_chem_bp_cols = np.loadtxt('results_2023_10_18/selected_chem_bp_cols.txt', dtype=str, delimiter='\t')
selected_prot_cols = np.loadtxt('results_2023_10_18/selected_prot_cols.txt', dtype=str, delimiter='\t')
selected_met_cols = np.loadtxt('results_2023_10_18/selected_met_cols.txt', dtype=str, delimiter='\t')

#  try taking only the first patient?
first_items = []
prev_client = None
for c in combined_table.public_client_id:
    if c == prev_client:
        first_items.append(False)
    else:
        prev_client = c
        first_items.append(True)
first_items = np.array(first_items)

combined_table_unique_subjects = combined_table[first_items]

## 2. Do differential expression for proteins and metabolites and clinical data based on delta-HbA1C

import scipy
import numpy as np

def data_test(data1, data2, test='u', use_fdr=True):
    """
    data1 and data2 are dataframes with the same columns
    Assume that they have the same columns
    NaNs are ignored?
    Do we assume that the data is already log-transformed?
    """
    data1_means = data1.mean(0)
    data2_means = data2.mean(0)
    mean_diff = data1_means - data2_means
    if test == 'u':
        t_test_results = scipy.stats.mannwhitneyu(data1, data2, axis=0, alternative='two-sided')
    elif test == 'w':
        t_test_results = scipy.stats.ranksums(data1, data2, axis=0, alternative='two-sided')
    elif test == 't':
        t_test_results = scipy.stats.ttest_ind(data1, data2, axis=0, alternative='two-sided')
    pvals = t_test_results.pvalue
    if use_fdr:
        pvals = scipy.stats.false_discovery_control(pvals, method='by')
    return mean_diff, pvals

import random

def get_control_set(normal_table, test_table):
    """
    Gets similar patients based on demographic characteristics
    """
    control_rows = []
    included_indices = set()
    all_similar_indices = set()
    for i, row in test_table.iterrows():
        age = row.age
        is_m = row.is_m
        min_age = age - 2
        max_age = age + 2
        similar_rows = normal_table[(normal_table.is_m == is_m) & (normal_table.age >= min_age) & (normal_table.age <= max_age)]
        indices = similar_rows.index.to_list()
        indices = [x for x in indices if x not in included_indices]
        try:
            sample = random.sample(indices, 2)
            included_indices.update(sample)
        except(ValueError):
            print('unable to find matching sample', age, is_m)
            min_age = age - 10
            max_age = age + 10
            similar_rows = normal_table[(normal_table.is_m == is_m) & (normal_table.age >= min_age) & (normal_table.age <= max_age)]
            indices = similar_rows.index.to_list()
            indices = [x for x in indices if x not in included_indices]
            if len(indices) == 1:
                sample = indices
            elif len(indices) >= 2:
                sample = random.sample(indices, 2)
            else:
                print('no matching sample found')
                sample = []
            included_indices.update(sample)
    return included_indices

### Iterating through different clinical variables and thresholds to do tests

def get_diffexp(data, variable, high_threshold, low_threshold,
                selected_prots=selected_prot_cols, selected_mets=selected_met_cols,
                selected_clinical=selected_chem_bp_cols):
    "Compute differential expression"
    data_high = data[data[variable] >= high_threshold]
    print('number of data points for {0}:'.format(variable), len(data_high))
    data_low = data[data[variable] < low_threshold]
    control_indices = get_control_set(data_low, data_high)
    data_control = data_low.loc[list(control_indices)]
    mean_diff, pvals = data_test(data_high[selected_prots], data_control[selected_prots], test='u')
    pvals = pd.Series(pvals, index=mean_diff.index)
    print('number of significant proteins for {0}:'.format(variable), (pvals < 0.05).sum())
    mets_mean_diff, mets_pvals = data_test(data_high[selected_mets], data_control[selected_mets], test='u')
    mets_pvals = pd.Series(mets_pvals, index=mets_mean_diff.index)
    print('number of significant metabolites for {0}:'.format(variable), (mets_pvals < 0.05).sum())
    chems_mean_diff, chems_pvals = data_test(data_high[selected_clinical], data_control[selected_clinical], test='u')
    chems_pvals = pd.Series(chems_pvals, index=chems_mean_diff.index)
    print('number of significant clinical tests for {0}:'.format(variable), (chems_pvals < 0.05).sum())
    return mean_diff, pvals, mets_mean_diff, mets_pvals, chems_mean_diff, chems_pvals

def get_diffexp_class(data, variable,
                selected_prots=selected_prot_cols, selected_mets=selected_met_cols,
                selected_clinical=selected_chem_bp_cols):
    "Compute differential expression based on a binarized class variable"
    data_high = data[data[variable]==1]
    print('number of data points for {0}:'.format(variable), len(data_high))
    data_low = data[data[variable]==0]
    control_indices = get_control_set(data_low, data_high)
    data_control = data_low.loc[list(control_indices)]
    mean_diff, pvals = data_test(data_high[selected_prots], data_control[selected_prots], test='u')
    pvals = pd.Series(pvals, index=mean_diff.index)
    print('number of significant proteins for {0}:'.format(variable), (pvals < 0.05).sum())
    mets_mean_diff, mets_pvals = data_test(data_high[selected_mets], data_control[selected_mets], test='u')
    mets_pvals = pd.Series(mets_pvals, index=mets_mean_diff.index)
    print('number of significant metabolites for {0}:'.format(variable), (mets_pvals < 0.05).sum())
    chems_mean_diff, chems_pvals = data_test(data_high[selected_clinical], data_control[selected_clinical], test='u')
    chems_pvals = pd.Series(chems_pvals, index=chems_mean_diff.index)
    print('number of significant clinical tests for {0}:'.format(variable), (chems_pvals < 0.05).sum())
    return mean_diff, pvals, mets_mean_diff, mets_pvals, chems_mean_diff, chems_pvals

variables = ['d_HbA1C_class', 'd_GLUCOSE_class', 'd_INSULIN_class', 'd_HOMA-IR_class', 'd_GFR_class']
    
# current values - TODO: in the previous analyses, these used all of the tables, not just the data subset. That's not done here.
base_variables = ['GLYCOHEMOGLOBIN A1C', 'GLUCOSE']
base_high_thresholds = [6.5, 125]
base_low_thresholds = [6.5, 125]

prots_mean_diff = []
prots_pvals = []
mets_mean_diff = []
mets_pvals = []
chems_mean_diff = []
chems_pvals = []
diffexp_outputs = []
base_diffexp_outputs = []
for var in variables:
    diffexp_output = get_diffexp_class(combined_table_unique_subjects, var)
    diffexp_outputs.append(diffexp_output)
for var, low, high in zip(base_variables, base_low_thresholds, base_high_thresholds):
    diffexp_output = get_diffexp(combined_table_unique_subjects,
                                                    var, high, low)
    base_diffexp_outputs.append(diffexp_output)

### Exporting protein and metabolite tables

prot_names = pd.read_csv('../arivale_data/arivale_prots.tsv', sep='\t')
met_names = pd.read_csv('../arivale_data/arivale_mets.tsv', sep='\t')

met_names.index = met_names.CHEMICAL_ID
prot_names.index = prot_names['index']

top_prots = prot_names.copy()
top_mets = met_names.copy()

prot_new_cols = []
met_new_cols = []
for var, diffexp_output in zip(variables, diffexp_outputs):
    column1_name = 'positive_' + var + '_mean_diff'
    column2_name = 'positive_' + var + '_pvals'
    top_prots[column1_name] = diffexp_output[0]
    top_prots[column2_name] = diffexp_output[1]
    prot_new_cols += [column1_name, column2_name]
    diffexp_output[2].index = diffexp_output[2].index.astype(int)
    diffexp_output[3].index = diffexp_output[3].index.astype(int)
    top_mets[column1_name] = diffexp_output[2]
    top_mets[column2_name] = diffexp_output[3]
    met_new_cols += [column1_name, column2_name]
    
# TODO: add current variables too
for var, diffexp_output in zip(base_variables, base_diffexp_outputs):
    column1_name = 'high_' + var + '_mean_diff'
    column2_name = 'high_' + var + '_pvals'
    top_prots[column1_name] = diffexp_output[0]
    top_prots[column2_name] = diffexp_output[1]
    prot_new_cols += [column1_name, column2_name]
    diffexp_output[2].index = diffexp_output[2].index.astype(int)
    diffexp_output[3].index = diffexp_output[3].index.astype(int)
    top_mets[column1_name] = diffexp_output[2]
    top_mets[column2_name] = diffexp_output[3]
    met_new_cols += [column1_name, column2_name]

top_prots = top_prots[['name', 'uniprot', 'gene_name'] + prot_new_cols]

top_prots = top_prots[~top_prots.positive_d_HbA1C_class_mean_diff.isna()]

top_prots.to_csv('results_2023_10_18/prots_diffexp_d_hba1c.csv', index=None)

top_mets = top_mets[['index', 'CHEMBL', 'KEGG', 'PUBCHEM'] + met_new_cols]

top_mets = top_mets[~top_mets['positive_d_HbA1C_class_mean_diff'].isna()]

top_mets.to_csv('results_2023_10_18/mets_diffexp_d_hba1c.csv', index=None)

### Export clinical table

clinical_columns = {}
for var, diffexp_output in zip(variables, diffexp_outputs):
    column1_name = 'positive_' + var + '_mean_diff'
    column2_name = 'positive_' + var + '_pvals'
    clinical_columns[column1_name] = diffexp_output[4]
    clinical_columns[column2_name] = diffexp_output[5]
    
for var, diffexp_output in zip(base_variables, base_diffexp_outputs):
    column1_name = 'high_' + var + '_mean_diff'
    column2_name = 'high_' + var + '_pvals'
    clinical_columns[column1_name] = diffexp_output[4]
    clinical_columns[column2_name] = diffexp_output[5]

clinical_diffexp = pd.DataFrame(clinical_columns)

clinical_diffexp.sort_values('positive_d_HbA1C_class_pvals')

clinical_diffexp.to_csv('results_2023_10_18/clinical_diffexp_d_hba1c.csv')
